backend/requirements.txt
pvporcupine
pyaudio
google-genai
elevenlabs
python-dotenv
webrtcvad-wheels
faster-whisper
numpy
asyncpg
